Boundary avoidance: first iteration
Previous: use Wrap()
Now: boids should avoid boundaries now

I had to increase the size of the boundary a little bit so that some of the boids won't escape in the beginning of the simulation.

I also noticed that some boids would clutter in the corners when separation factor is high. But if I increased the cohesion factor and then decreased it, some boids would go back to normal